backend/requirements.txt
django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
pymongo>=4.0
mongoengine>=0.27
django-cors-headers>=4.0
channels>=4.0
